Major Rust refactoring
- Now glue is generated into crates
- Uses #[export()] attribute which type checks and asserts export exists
- This depends on bento-macros, which due to idiosyncrasies must be its
  own crate. Most likely this crate will be submitted to crates.io in
  the future.
- Also added a multilang examples

Unfortunately I am noticing a significant increase in size with Rust.
Debug builds are nearly unusable (may just require change in development
style).

Here is a release build of the multilang example, which has quick-sort
implemented in both C and Rust:

   text    data     bss     dec     hex filename
   4916       8     116    5040    13b0 sys.elf
   1252       0      92    1344     540 boxc.elf
   4324       0      92    4416    1140 boxrust.elf
  10484       8     300   10792    2a28 (TOTALS)

It may be that a lot of the overhead is coming from the C-FFI
conversion, there may be room for improvement there.