Fixed several issues with Wamr, added support for rest of examples
See extras/wamr-modified for the local modifications, there were a
number needed to get things working, especially without running out
of memory due to excessive fragmentation when we page in and out boxes.

The most hacky change is to cache the most recently allocated "large" block
of memory. This helps blocks survive paging in and out of init state when
memory is really tight, but really is a hacky hack. Also Wamr's biggest
allocation is 64*1024+36, which is a really annoying number for the
allocator...